+++ src/support/elements/definitions/index.js
@@ -33,6 +33,8 @@ export {
 export {
   buildCitedDomainsPayload, transformCitedDomainsResponse, transformCitedDomainsResponses,
 } from './cited-domains.js';
+export { buildSubredditsPayload, transformSubredditsResponse } from './subreddits.js';
+export { buildRedditThreadsPayload, transformRedditThreadsResponse } from './reddit-threads.js';
 export { buildTopicPromptsPayload, transformTopicPromptsResponse } from './topic-prompts.js';
 export { aggregateTopicsFromPrompts } from './topics-insights.js';
 export {